Disable share button on edit page
