+++ app/src/marketplace/components/StreamPreviewPage/index.jsx
@@ -22,7 +22,7 @@ import Notification from '$shared/utils/Notification'
 import { NotificationIcon } from '$shared/utils/constants'
 import { Context as ClientContext, Provider as ClientProvider } from '$shared/contexts/StreamrClient'
 import useIsMounted from '$shared/hooks/useIsMounted'
-import useStreamPermissions from '$shared/hooks/useStreamPermissions'
+import useStreamPermissions from '$userpages/hooks/useStreamPermissions'
 import ResourceNotFoundError, { ResourceType } from '$shared/errors/ResourceNotFoundError'
 
 type Props = {
