--- app/src/shared/hooks/useStreamPermissions.js
@@ -1,66 +1,0 @@
-// @flow
-
-import { useState, useEffect } from 'react'
-import { useSelector } from 'react-redux'
-import { getResourcePermissions } from '$userpages/modules/permission/services'
-import { canHandleLoadError, handleLoadError } from '$auth/utils/loginInterceptor'
-import ResourceNotFoundError from '$shared/errors/ResourceNotFoundError'
-import useFailure from '$shared/hooks/useFailure'
-import { selectUserData, isAuthenticating } from '$shared/modules/user/selectors'
-import useIsMounted from '$shared/hooks/useIsMounted'
-
-const extractOperations = (data: Array<any>, username: ?string): Array<string> => (
-    data
-        .filter((item) => item.anonymous || (username && item.user === username))
-        .map((item) => item.operation)
-)
-
-export default (id: string) => {
-    const [permissions, setPermissions] = useState()
-
-    const fail = useFailure()
-
-    const { username } = useSelector(selectUserData) || {}
-
-    const authenticating = useSelector(isAuthenticating)
-
-    const isMounted = useIsMounted()
-
-    useEffect(() => {
-        setPermissions(undefined)
-
-        const fetch = async () => {
-            try {
-                try {
-                    const data = await getResourcePermissions({
-                        resourceType: 'STREAM',
-                        resourceId: id,
-                        id: 'me',
-                    })
-
-                    if (isMounted()) {
-                        setPermissions(extractOperations(data, username))
-                    }
-                } catch (error) {
-                    if (canHandleLoadError(error)) {
-                        await handleLoadError({
-                            error,
-                        })
-                    }
-                }
-            } catch (e) {
-                if (e instanceof ResourceNotFoundError) {
-                    fail(e)
-                    return
-                }
-                throw e
-            }
-        }
-
-        if (!authenticating) {
-            fetch()
-        }
-    }, [id, username, authenticating, fail, isMounted])
-
-    return permissions
-}
