+++ app/src/userpages/components/StreamPage/Edit/edit.pcss
@@ -1,14 +1,3 @@
-.longText {
-  max-width: 536px;
-}
-
-.statusTitle {
-  display: inline-flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-}
-
 .showOnDesktop {
   display: none;
 }
